requirements.txt
python-telegram-bot==13.15
Django==4.2.7
environs==9.5.0